Added market value version, Moved logic to static class
